CP-54351 move resetvdis from /opt
